Code improvement with C++ 11

+++ src/torrentlistctrl.cpp
@@ -138,7 +138,7 @@ wxString TorrentListCtrl::GetItemValue( long item, long columnid ) const
 		_( "Allocating" ),
 		_( "Checking resume data" )
 	};
-	static unsigned int state_size = ( sizeof( state_str ) / ( sizeof( wxChar * ) ) );
+	static size_t state_size = ( sizeof( state_str ) / ( sizeof( wxChar * ) ) );
 
 	wxString ret;
